购物车crud，远程调用，feign远程调用流程

+++ model/src/main/java/com/atguigu/gmall/product/entity/SkuInfo.java
@@ -33,7 +33,7 @@ public class SkuInfo implements Serializable {
      * 价格
      */
     @TableField(value = "price")
-    private Integer price;
+    private BigDecimal price;
 
     /**
      * sku名称
